removed compiler warning

+++ rost_vision/src/flann_matcher.hpp
@@ -43,7 +43,7 @@ public:
   }
 
   void get_words(cv::Mat q, vector<int>& words){
-    assert(q.cols == data.cols);
+    assert(static_cast<int>(q.cols) == static_cast<int>(data.cols));
     query = flann::Matrix<unsigned char>(q.data, q.rows, q.cols, q.step[0]);
     vector<DistanceType> dists_vec(query.rows);
     dists = flann::Matrix<DistanceType>(&(dists_vec[0]), query.rows, 1);
